+++ core/src/test/resources/resolveLock-test/q1.sql
@@ -1,6 +1,6 @@
 select
   sum(C_ACCTBAL)
 from
-  tidb_resolveLock_test.CUSTOMER
+  CUSTOMER
 where
   C_ACCTBAL % 2 == 0