+++ core/src/test/resources/resolveLock-test/sum_account.sql
@@ -1,4 +1,4 @@
 select
   sum(C_ACCTBAL)
 from
-  tidb_resolveLock_test.CUSTOMER
+  CUSTOMER