+++ core/src/test/scala/com/pingcap/tispark/datasource/EdgeConditionSuite.scala
@@ -67,7 +67,8 @@ class EdgeConditionSuite extends BaseDataSourceTest("test_datasource_edge_condit
     testTiDBSelect(Seq(row1, row2, row3, row4))
   }
 
-  test("Write to table with one column (primary key + auto increase)") {
+  // currently user provided auto increment value is not supported!
+  ignore("Write to table with one column (primary key + auto increase)") {
     if (!supportBatchWrite) {
       cancel
     }